Bugfixes, documentation

+++ src/scrapers/mongodb_handler.py
@@ -81,6 +81,17 @@ class MongoDBHandler(DBHandlerInterface):
         return True
 
     def rename_collection(self, collection_new_name: str, db_name: str = None, collection_name: str = None, **kwargs) -> bool:
+        """
+        Rename collection.
+
+        Args:
+            collection_new_name (str): New name of the collection
+            db_name (str, optional): Name of the database where the collection is. Defaults to None.
+            collection_name (str, optional): Name of the collection which is to be used. Defaults to None.
+
+        Returns:
+            bool: [description]
+        """
         db: Database = self.db if db_name is None else self.client[db_name]
         coll: Collection = self.coll if collection_name is None else db[collection_name]
 
@@ -89,9 +100,19 @@ class MongoDBHandler(DBHandlerInterface):
         return True
 
     def drop_collection(self, db_name: str = None, collection_name: str = None, **kwargs) -> bool:
+        """
+        Drop collection if it exists.
+
+        Args:
+            db_name (str, optional): Name of the database where the collection is. Defaults to None.
+            collection_name (str, optional): Name of the collection which is to be used. Defaults to None.
+
+        Returns:
+            bool: [description]
+        """
         db: Database = self.db if db_name is None else self.client[db_name]
         coll: Collection = self.coll if collection_name is None else db[collection_name]
-
-        coll.drop()
+        if(coll.name in db.list_collection_names()):
+            coll.drop()
 
         return True
